requirements.txt
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
idna==2.10
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==3.2.1
numpy==1.18.2
pandas==1.0.4
pkg-resources==0.0.0
progressbar2==3.50.1
pymongo==3.5.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2020.1
requests==2.25.1
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
terminaltables==3.1.0
urllib3==1.26.3
